CAMBIOS

+++ PotaxieSport/Models/MovimientoEconomico.cs
@@ -7,9 +7,10 @@
         public int ContadorId { get; set; }
         public string? Contador { get; set; }
         public string? Tipo { get; set; }
-        public Double Cantidad { get; set; }
+        public Decimal Cantidad { get; set; }
         public int TorneoId { get; set; }
         public string? Torneo { get; set; }
         public string? Comprobante { get; set; }
+        
     }
 }
